+++ types.ts
@@ -2,9 +2,9 @@ export interface AdapterOptions {
   /**
    * OpenAI API key. Go to https://platform.openai.com/, go to Dashboard -> API keys -> Create new secret key
    * Paste value in your .env file OPENAI_API_KEY=your_key
-   * Set openAiApiKey: process.env.OPENAI_API_KEY to access it
+   * Set geminiApiKey: process.env.GEMINI_API_KEY to access it
    */
-  openAiApiKey: string;
+  geminiApiKey: string;
 
   /**
    * Model name. Go to https://platform.openai.com/docs/models, select model and copy name.